refactor: move methods to slack service for reusability
