+++ src/modules/playlist/digest.service.ts
@@ -1,9 +1,8 @@
 import { SlackService } from '@modules/slack';
-import { TrackSourceProvider } from '@modules/track/enums';
+import { MessageWithMusicSummery } from '@modules/slack/types';
 import { Track } from '@modules/track/track.entity';
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
-import urlParser from 'js-video-url-parser';
 import { PageRequest, PageResponse } from 'src/interface-adapters/dtos';
 import {
   CreateDigestRequest,
@@ -13,15 +12,6 @@ import {
 import { PlaylistType } from './enums';
 import { Playlist } from './playlist.entity';
 import { PlaylistRepository } from './playlist.repository';
-
-export interface MessageSummery {
-  author: string;
-  text: string;
-  musicService: string;
-  musicUrl: string;
-  musicTitle: string;
-  reactionCount: number;
-}
 
 @Injectable()
 export class DigestService {
@@ -89,35 +79,15 @@ export class DigestService {
 
     const memberIdMap = await this.slackService.getMemberSlackIdMap();
     const messagesWithMusic = messages.reduce((total, message) => {
-      const initialMusic = message.attachments?.find(
-        (it) =>
-          it.service_name &&
-          ['YouTube', 'SoundCloud'].includes(it.service_name) &&
-          this.isSupportedMusicUrl(it.original_url),
+      const summary = this.slackService.summarizeMessageWithMusicUsingMap(
+        message,
+        memberIdMap,
       );
-      if (!initialMusic) {
+      if (!summary) {
         return total;
       }
-
-      const text =
-        message.blocks
-          ?.find((it) => it.type === 'rich_text')
-          ?.elements?.find((it) => it.type === 'rich_text_section')
-          // @ts-ignore: Element.elements가 존재하는데 타이핑이 잘못됨
-          ?.elements?.find((it) => it.type === 'text')
-          ?.text?.trim() ?? '';
-      const reactionCount =
-        message.reactions?.reduce((sum, cur) => sum + (cur.count ?? 0), 0) ?? 0;
-      const summary: MessageSummery = {
-        author: memberIdMap.get(message.user ?? '')?.name ?? 'undefined',
-        text: text,
-        musicService: initialMusic.service_name!,
-        musicUrl: this.hydrateSupportedMusicUrl(initialMusic.original_url),
-        musicTitle: initialMusic.title!,
-        reactionCount,
-      };
       return [...total, summary];
-    }, [] as MessageSummery[]);
+    }, [] as MessageWithMusicSummery[]);
     const topMessages = messagesWithMusic
       .sort((a, b) => b.reactionCount - a.reactionCount)
       .slice(0, 10);
@@ -129,7 +99,9 @@ export class DigestService {
     digest.tracks.set(
       topMessages.map((it, index) => {
         const track = new Track();
-        track.sourceProvider = this.mapServiceProvider(it.musicService);
+        track.sourceProvider = this.slackService.mapServiceProvider(
+          it.musicService,
+        );
         track.sourceUrl = it.musicUrl;
         track.position = index + 1;
         track.title = it.musicTitle;
@@ -140,42 +112,5 @@ export class DigestService {
     );
     await this.playlistRepo.persistAndFlush(digest);
     return DigestResponse.fromEntity(digest);
-  }
-
-  private isSupportedMusicUrl(url: string | undefined) {
-    if (!url) {
-      return false;
-    }
-    const parsed = urlParser.parse(url);
-    if (!parsed) {
-      return false;
-    }
-    if (!['video', 'track'].includes(parsed.mediaType || '')) {
-      return false;
-    }
-    return true;
-  }
-
-  private hydrateSupportedMusicUrl(url: string | undefined) {
-    if (!this.isSupportedMusicUrl(url)) {
-      throw new Error('Not supported music url');
-    }
-    const hydrated = urlParser.create({ videoInfo: urlParser.parse(url!)! });
-    if (!hydrated) {
-      throw new Error('Cannot hydrate url');
-    }
-    return hydrated;
-  }
-
-  private mapServiceProvider(name: string) {
-    const map: Record<string, TrackSourceProvider> = {
-      YouTube: TrackSourceProvider.YOUTUBE,
-      SoundCloud: TrackSourceProvider.SOUNDCLOUD,
-    };
-    const provider = map[name];
-    if (!provider) {
-      throw new Error('Cannot map service provider');
-    }
-    return provider;
   }
 }
